+++ platformio/package/manager/_install.py
@@ -20,8 +20,8 @@ import tempfile
 import click
 
 from platformio import app, compat, fs, util
-from platformio.package.exception import PackageException
-from platformio.package.meta import PackageItem
+from platformio.package.exception import PackageException, UnknownPackageError
+from platformio.package.meta import PackageItem, PackageSpec
 from platformio.package.unpack import FileUnpacker
 from platformio.package.vcsclient import VCSClientFactory
 
@@ -115,7 +115,38 @@ class PackageManagerInstallMixin(object):
         return pkg
 
     def install_dependencies(self, pkg):
-        pass
+        assert isinstance(pkg, PackageItem)
+        manifest = self.load_manifest(pkg)
+        dependencies = manifest.get("dependencies")
+        if not dependencies:
+            return
+        self.log.info("Resolving dependencies...")
+        for dependency in dependencies:
+            if not self._install_dependency(dependency):
+                self.log.warning(
+                    click.style(
+                        "Warning! Could not install dependency %s for package '%s'"
+                        % (dependency, pkg.metadata.name),
+                        fg="yellow",
+                    )
+                )
+
+    def _install_dependency(self, dependency):
+        spec = PackageSpec(
+            owner=dependency.get("owner"),
+            name=dependency.get("name"),
+            requirements=dependency.get("version"),
+        )
+        search_filters = {
+            key: value
+            for key, value in dependency.items()
+            if key in ("authors", "platforms", "frameworks")
+        }
+        try:
+            return self._install(spec, search_filters=search_filters or None)
+        except UnknownPackageError:
+            pass
+        return None
 
     def install_from_url(self, url, spec, checksum=None):
         spec = self.ensure_spec(spec)