+++ platformio/package/manager/_uninstall.py
@@ -19,7 +19,7 @@ import click
 
 from platformio import fs
 from platformio.package.exception import UnknownPackageError
-from platformio.package.meta import PackageSpec
+from platformio.package.meta import PackageItem, PackageSpec
 
 
 class PackageManagerUninstallMixin(object):
@@ -75,4 +75,18 @@ class PackageManagerUninstallMixin(object):
         return pkg
 
     def uninstall_dependencies(self, pkg):
-        pass
+        assert isinstance(pkg, PackageItem)
+        manifest = self.load_manifest(pkg)
+        if not manifest.get("dependencies"):
+            return
+        self.log.info(click.style("Removing dependencies...", fg="yellow"))
+        for dependency in manifest.get("dependencies"):
+            spec = PackageSpec(
+                owner=dependency.get("owner"),
+                name=dependency.get("name"),
+                requirements=dependency.get("version"),
+            )
+            pkg = self.get_package(spec)
+            if not pkg:
+                continue
+            self._uninstall(pkg)